+++ scripts/build.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 set -ex
 
 # Optional flags
@@ -17,7 +17,7 @@ python3 -m venv "$venv_dir"
 export PATH="$venv_dir/bin:$PATH"
 
 # Remove old build
-rm -v dist/*
+rm -v dist/* || true
 # Install requirements
 pip install -r requirements.txt
 # Build package
@@ -52,7 +52,7 @@ done
 if [ $install_pyo3_build -gt 0 ]; then
   pip install -r requirements.txt
   bash scripts/ci_install_pyo3_builds.sh
-  pip install mypy types-termcolor>=1.1 types_redis>=4.3 requests>=2.28 types-requests>=2.28
+  pip install mypy "types-termcolor>=1.1" "types_redis>=4.3" "requests>=2.28" "types-requests>=2.28"
 fi
 if [ $test -gt 0 ]; then
   pip install dist/cloudproof_py*.whl