+++ scripts/ci_install_pyo3_builds.sh
@@ -8,9 +8,9 @@ install_lib() {
         rm linux.zip && rm -rf x86_64*
 }
 
-install_lib "v2.4.0"
+install_lib "v3.0.0"
 if [ $? -ne 0 ]; then
-    branch="fix/rename_findex_traits"
+    branch="feature/covercrypt_rekey"
     echo installing python from branch ${branch}
     install_lib "last_build/${branch}"
 fi